requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7

# LLM API clients
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2

# Vector database and embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# AST analysis
ast-comments>=0.1.0
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4

# Code analysis and formatting
black>=24.3.0
ruff>=0.3.4
pylint>=3.1.0
astunparse>=1.6.3

# File and repository handling
gitpython>=3.1.40
pathspec>=0.12.1

# Additional analysis tools
networkx>=3.2.1  # For dependency graph visualization

# Utilities
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0

